+++ DependencyInjection/DeveloptechAgilityExtension.php
@@ -24,6 +24,7 @@ class DeveloptechAgilityExtension extends Extension
 
         $container->setParameter('developtech_agility.model.project.class', $config['model']['class']['project']);
         $container->setParameter('developtech_agility.model.feature.class', $config['model']['class']['feature']);
+        $container->setParameter('developtech_agility.model.feedback.class', $config['model']['class']['feedback']);
 
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yml');