+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [Unreleased]
+### Added
+* Feedback model
+* Project feedbacks list service
+* Get feedback service
+* Feedback creation service
+
 ## [0.1.0] - 2016-07-13
 ### Added
 * Project model